+++ src/main/java/hanuri/website/model/ImageSource.java
@@ -3,7 +3,7 @@ package hanuri.website.model;
 import hanuri.website.dto.image.EImageType;
 
 public interface ImageSource {
-    int getObjectId();
+    Long getObjectId();
 
     EImageType getImageType();
 }
